+++ Tests/JExtractSwiftTests/JNI/JNIGenericTypeTests.swift
@@ -120,7 +120,7 @@ struct JNIGenericTypeTests {
         extension MyID: _SwiftModule_MyID_opener {
           static func _get_description(environment: UnsafeMutablePointer<JNIEnv?>!, thisClass: jclass, selfPointer: jlong) -> jstring? {
             assert(selfPointer != 0, "selfPointer memory address was null")
-            let selfPointerBits$ = Int(Int64(fromJNI: selfPointer, in: environment))
+            let selfPointerBits$ = Int(fromJNI: selfPointer, in: environment)
             let selfPointer$ = UnsafeMutablePointer<MyID>(bitPattern: selfPointerBits$)
             guard let selfPointer$ else {
               fatalError("selfPointer memory address was null in call to \(#function)!")
@@ -133,7 +133,7 @@ struct JNIGenericTypeTests {
         """
         @_cdecl("Java_com_example_swift_MyID__00024getDescription__JJ")
         public func Java_com_example_swift_MyID__00024getDescription__JJ(environment: UnsafeMutablePointer<JNIEnv?>!, thisClass: jclass, selfPointer: jlong, selfTypePointer: jlong) -> jstring? {
-          let selfTypePointerBits$ = Int(Int64(fromJNI: selfTypePointer, in: environment))
+          let selfTypePointerBits$ = Int(fromJNI: selfTypePointer, in: environment)
           guard let selfTypePointer$ = UnsafeRawPointer(bitPattern: selfTypePointerBits$) else {
             fatalError("selfTypePointer metadata address was null")
           }
@@ -188,11 +188,11 @@ struct JNIGenericTypeTests {
         public func Java_com_example_swift_SwiftModule__00024makeStringID__Ljava_lang_String_2Lorg_swift_swiftkit_core__1OutSwiftGenericInstance_2(environment: UnsafeMutablePointer<JNIEnv?>!, thisClass: jclass, value: jstring?, resultOut: jobject?) {
           let result$ = UnsafeMutablePointer<MyID<String>>.allocate(capacity: 1)
           result$.initialize(to: SwiftModule.makeStringID(String(fromJNI: value, in: environment)))
-          let resultBits$ = Int64(Int(bitPattern: result$))
+          let resultBits$ = Int(bitPattern: result$)
           do {
             environment.interface.SetLongField(environment, resultOut, _JNIMethodIDCache._OutSwiftGenericInstance.selfPointer, resultBits$.getJNIValue(in: environment))
             let metadataPointer = unsafeBitCast(MyID<String>.self, to: UnsafeRawPointer.self)
-            let metadataPointerBits$ = Int64(Int(bitPattern: metadataPointer))
+            let metadataPointerBits$ = Int(bitPattern: metadataPointer)
             environment.interface.SetLongField(environment, resultOut, _JNIMethodIDCache._OutSwiftGenericInstance.selfTypePointer, metadataPointerBits$.getJNIValue(in: environment))
           }
           return
@@ -202,7 +202,7 @@ struct JNIGenericTypeTests {
         @_cdecl("Java_com_example_swift_SwiftModule__00024takeIntID__J")
         public func Java_com_example_swift_SwiftModule__00024takeIntID__J(environment: UnsafeMutablePointer<JNIEnv?>!, thisClass: jclass, value: jlong) -> jlong {
           assert(value != 0, "value memory address was null")
-          let valueBits$ = Int(Int64(fromJNI: value, in: environment))
+          let valueBits$ = Int(fromJNI: value, in: environment)
           let value$ = UnsafeMutablePointer<MyID<Int>>(bitPattern: valueBits$)
           guard let value$ else {
             fatalError("value memory address was null in call to \\(#function)!")
